Adjust visiblity on SmithyBuild* objects
