+++ modules/codegen/src/smithy4s/codegen/SmithyBuildJson.scala
@@ -18,26 +18,9 @@ package smithy4s.codegen
 
 import smithy4s.codegen.internals.SmithyBuild
 import smithy4s.codegen.internals.SmithyBuildMaven
+import upickle.default._
 
-private[codegen] final case class SmithyBuildData(
-    imports: Seq[String],
-    deps: Seq[String],
-    repos: Seq[String]
-) {
-  def addAll(
-      imports: Seq[String],
-      deps: Seq[String],
-      repos: Seq[String]
-  ): SmithyBuildData = {
-    SmithyBuildData(
-      this.imports ++ imports,
-      this.deps ++ deps,
-      this.repos ++ repos
-    )
-  }
-}
-
-object SmithyBuildJson {
+private[codegen] object SmithyBuildJson {
 
   def toJson(
       imports: Seq[String],
@@ -57,6 +40,46 @@ object SmithyBuildJson {
       json1: String,
       json2: String
   ): String = {
-    SmithyBuild.merge(json1, json2)
+    val j1 = read[ujson.Value](json1)
+    val j2 = read[ujson.Value](json2)
+    val merged = mergeJs(j1, j2)
+    val finalJs = removeArrayDuplicates(merged)
+    finalJs.render(indent = 4)
+  }
+
+  private def mergeJs(
+      v1: ujson.Value,
+      v2: ujson.Value
+  ): ujson.Value = {
+    (v1, v2) match {
+      case (ujson.Obj(obj1), ujson.Obj(obj2)) =>
+        val result = obj2.foldLeft(obj1.toMap) {
+          case (elements, (key, value2)) =>
+            val value = elements.get(key) match {
+              case None =>
+                value2
+              case Some(value1) =>
+                mergeJs(value1, value2)
+            }
+            elements.updated(key, value)
+        }
+        ujson.Obj.from(result)
+      case (arr1: ujson.Arr, arr2: ujson.Arr) =>
+        ujson.Arr(arr1.arr ++ arr2.arr)
+      case (_, _) => v1
+    }
+  }
+
+  private def removeArrayDuplicates(js: ujson.Value): ujson.Value = {
+    js match {
+      case ujson.Obj(obj1) =>
+        ujson.Obj.from(
+          obj1.toList.map { case (key, value) =>
+            key -> removeArrayDuplicates(value)
+          }
+        )
+      case (arr1: ujson.Arr) => arr1.arr.distinct
+      case x                 => x
+    }
   }
 }
